move min,max,clamp,clamp_float to misc
