+++ src/listbox.c
@@ -3,7 +3,6 @@
 #include <stdlib.h>
 #include "log.h"
 #include "misc.h"
-#include "linalg.h"
 
 void listbox_init(struct Listbox *lb)
 {
