+++ src/misc.h
@@ -5,6 +5,21 @@
 #include <SDL2/SDL_ttf.h>
 #include <stdint.h>
 
+// windows.h defines min and max macros just like the ones below
+#if !defined(min) && !defined(max)
+	#define min(a,b) ((a)<(b) ? (a) : (b))
+	#define max(a,b) ((a)>(b) ? (a) : (b))
+#endif
+
+inline void clamp(int *val, int lo, int hi) {
+	if (*val < lo) *val = lo;
+	if (*val > hi) *val = hi;
+}
+inline void clamp_float(float *val, float lo, float hi) {
+	if (*val < lo) *val = lo;
+	if (*val > hi) *val = hi;
+}
+
 enum MiscState {
 	MISC_STATE_CHOOSER,
 	MISC_STATE_MAPEDITOR,
